Big update: Convert Stylish to a framework target
- Link Stylish in as a framework rather than a file to copy and paste
- Updated handling of JSONStyleProperties so clinet apps can create and add their own
- Made global vars for cutom dynamic property sets and shared style classes, so it doesn't require subclassing and overriding to use them
- StylishExample project updated to link in Stylish as a subproject and framework
- Other small updates and improvements
