+++ Stylish.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:StylishExample.xcodeproj">
+      location = "self:Stylish.xcodeproj">
    </FileRef>
 </Workspace>
