+++ Stylish/Stylish.h
@@ -1,0 +1,19 @@
+//
+//  Stylish.h
+//  Stylish
+//
+//  Created by Hall, Daniel on 8/11/17.
+//
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for Stylish.
+FOUNDATION_EXPORT double StylishVersionNumber;
+
+//! Project version string for Stylish.
+FOUNDATION_EXPORT const unsigned char StylishVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Stylish/PublicHeader.h>
+
+
