+++ StylishExample/Stylish/Stylish
@@ -1,0 +1,1 @@
+../../Stylish
