+++ StylishExample/Stylish/Stylish.xcodeproj
@@ -1,0 +1,1 @@
+../../Stylish.xcodeproj
