+++ StylishExample/StylishExample/AppDelegate.swift
@@ -2,40 +2,36 @@
 //  AppDelegate.swift
 //  StylishExample
 //
-// Copyright (c) 2016 Daniel Hall
-// Twitter: @_danielhall
-// GitHub: https://github.com/daniel-hall
-// Website: http://danielhall.io
+//  Created by Hall, Daniel on 8/11/17.
+//  Copyright © 2017 Hall, Daniel. All rights reserved.
 //
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-
 
 import UIKit
+import Stylish
 
+extension UIView {
+    // This is the only entry point for setting global variables in Stylish for Interface Builder rendering (since App Delegate doesn't get run by IBDesignable. So we are setting up the same global variable here so they are also available during live previews on Storyboards
+    open override func prepareForInterfaceBuilder() {
+        // Set our custom property sets that should be part of JSON parsing
+        Stylish.customDynamicPropertySets = [ProgressBarPropertySet.self]
+        // Set the style classes we want to be part of every stylesheet
+        Stylish.sharedStyleClasses = [("Rounded", RoundedStyle()), ("HighlightedText", HighlightedTextStyle())]
+    }
+}
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
 
+
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+        
+        // Set our custom property sets that should be part of JSON parsing
+        Stylish.customDynamicPropertySets = [ProgressBarPropertySet.self]
+        
+        // Set the style classes we want to be part of every stylesheet
+        Stylish.sharedStyleClasses = [("Rounded", RoundedStyle()), ("HighlightedText", HighlightedTextStyle())]
         
         //Remove the following code if you want to dynamically download and cache a stylesheet from the web.
         let paths = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true)
@@ -51,3 +47,4 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         return true
     }
 }
+
