+++ StylishExample/StylishExample/Assets.xcassets/Contents.json
