+++ StylishExample/StylishExample/Assets.xcassets/stone.imageset/Contents.json
