+++ StylishExample/StylishExample/Assets.xcassets/water.imageset/Contents.json
