--- StylishExample/StylishExampleTests/StylishExampleTests.swift
@@ -1,36 +1,0 @@
-//
-//  StylishExampleTests.swift
-//  StylishExampleTests
-//
-//  Created by Daniel Hall on 3/20/16.
-//  Copyright © 2016 Daniel Hall. All rights reserved.
-//
-
-import XCTest
-@testable import StylishExample
-
-class StylishExampleTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
-}
